added initial build system
